+++ services/java/ticket-order-api/src/main/java/com/example/ticketplatform/api/application/service/LoginService.java
@@ -3,7 +3,7 @@ package com.example.ticketplatform.api.application.service;
 import com.example.ticketplatform.api.application.port.in.LoginUseCase;
 import com.example.ticketplatform.api.application.port.in.LoginCommand;
 import com.example.ticketplatform.api.application.port.out.PasswordMatcherPort;
-import com.example.ticketplatform.api.application.port.out.UserRepositoryPort;
+import com.example.ticketplatform.api.application.port.out.UserAuthRepositoryPort;
 import com.example.ticketplatform.api.domain.model.user.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.BadCredentialsException;
@@ -15,13 +15,13 @@ import static java.lang.Boolean.TRUE;
 @RequiredArgsConstructor
 class LoginService implements LoginUseCase {
 
-  private final UserRepositoryPort userRepositoryPort;
+  private final UserAuthRepositoryPort userAuthRepositoryPort;
   private final PasswordMatcherPort passwordMatcherPort;
 
   @Override
   public User login(LoginCommand command) {
     User user =
-        userRepositoryPort
+        userAuthRepositoryPort
             .findByEmail(command.login())
             .orElseThrow(() -> new BadCredentialsException("Invalid credentials"));